Pipeline/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
sentence-transformers
pandas==2.3.3
# Core dependencies
tiktoken==0.5.2
python-dotenv==1.0.0
fastapi
# --- LANGCHAIN (Unpinned to prevent conflicts) ---
langchain==1.0.5
langchain-core==1.0.4
langchain-chroma==1.0.0
langchain-huggingface==1.0.1
langchain-anthropic==1.0.3

openai
anthropic
groq

# Neo4j and Graph
neo4j==6.0.3

networkx==3.5

# Database
supabase==2.16.0


# Embeddings and RAG
bm25s
scikit-learn

# Utilities
pydantic== 2.11.7
requests==2.32.4
# Visualization
matplotlib
numpy
